+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Before every entry put one of these to mark the severity of the change:
 ## Unreleased
 ### Added
 - [Minor] Added modbus flavour feature e.g. to specify device specific options.
+- [Minor] Added public API methods to EDrivePositioning for configuring various options (SINAPOS functionality).
 ## [v0.5.1] - 10.08.22
 ### Added
 - [Patch] Added proper README, docu and examples.